Add istanbul for code coverage
Instead of calling mocha directly the use
of nyc starts istanbul and the tests.

After the test have run, coverage reports
in HTML can be found in the .nyc_output/coverage
folder.

nyc also prints a report directly to the command
line, which in most cases actually should be
sufficient - unless your tests run on a build server.