Exclude special files from coverage
This concerns all tests files, which must not be
covered. And especially the files within the
dependencies folder (currently one).

That folder contains dependencies that are going
to be mocked. Thus the actual lines within
those files are never going to be covered.

The functions here should only create objects
or functions and not contain any logic at all.