Refactor usecases to no longer return an Observable
Instead I am now using a callback. This seems
to be a cleaner approach since I can keep the
main entry point (index.js) very concise and
devoid of any logic. It simply calls
app.ask and app.ell based on which
callback was called and (in some cases) which
text the usecase passed as an argument to
the callback.

One downside: the usecase is doing usecasey stuff
like calling the api and mapping it to a parser
result. But it does also UI logic stuff like
creating answer phrases. For a more complicated
usecase this probably has to be split up into
a proper usecase and one class doing the UI logic
to keep inline with the single responsibility principle).

But since the usecase would be three lines only
in this specific case, splitting it seems to be
overkill for now.

NOTE: The usecase tests have been removed. Code
coverage will be restored with the next commit
that tests the response to the Assistant based
on different request situations.

I prefer a very understandable commit history
over perfect coverage with every commit which is
why I do not commit both steps at once.