Create the app response in the firebase function
This creates the map for all the intents possible (in
this case only two) with matching handler functions
for the intents and passes everything on to the
usecase for processing.

The index.js file provides a callback that only
contains app.tell() and app.ask() calls (plus some
logging).

This commit also contains a test that covers the
desired responses for the possible conditions
and by doing so also tests all usecase
methods alongside it. This way the code coverage which
was broken with the last commit is now restored
for the usecases.